WIP: Clean up

+++ lib/src/Rendering/Graphics.hpp
@@ -38,11 +38,11 @@ namespace fm {
         ~Graphics();
         void Clear(BUFFER_BIT = COLOR_BUFFER_BIT) const;
         void SetViewPort(const fm::math::vec4i &rect) const;
-        void SetViewPort(const fm::Rect<float> &rect) const;
+        void SetViewPort(const fm::Rect<int> &rect) const;
         void Enable(RENDERING_TYPE renderingTYPE) const;
         void Disable(RENDERING_TYPE renderingTYPE) const;
-        void Draw(int primitiveType, unsigned int vertexCount, unsigned int* indices) const;
-        void Draw(int primitiveType, unsigned int vertexStart, unsigned int vertexCount) const;
+        void Draw(int primitiveType, size_t vertexCount, size_t* indices) const;
+        void Draw(int primitiveType, size_t vertexStart, size_t vertexCount) const;
         void Draw(Model* model) const;
         void BindVertexBuffer(rendering::VertexBuffer *vertexBuffer) const;
         void BindFrameBuffer(unsigned int id) const;
